# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QA Camera Forms API",
    "version": "1.0.0",
    "description": "OpenAPI spec generated from the provided Laravel routes/controllers/models.\n\nNotes:\n- All JSON responses follow a wrapper: { status, message, data, errors }.\n- Auth is required for all endpoints (middleware AuthTokenStoreScopeMiddleware)."
  },
  "servers": [
    {
      "url": "https://qa.lcportal.cloud",
      "description": "Production"
    },
    {
      "url": "http://localhost",
      "description": "Local"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/camera-forms/{id}": {
      "get": {
        "tags": ["Camera Forms"],
        "summary": "Show a single audit (camera form bundle) by audit id",
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Camera form fetched successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiSuccessResponse" },
                "examples": {
                  "success": {
                    "value": {
                      "status": "success",
                      "message": "Camera form fetched successfully",
                      "data": {
                        "id": 91,
                        "store_id": 12,
                        "user_id": 5,
                        "date": "2026-01-20",
                        "store": {
                          "id": 12,
                          "store": "Brooklyn - Atlantic Ave",
                          "group": 3
                        },
                        "user": {
                          "id": 5,
                          "name": "Jane Auditor",
                          "email": "jane@example.com"
                        },
                        "camera_forms": [
                          {
                            "id": 210,
                            "user_id": 5,
                            "entity_id": 33,
                            "audit_id": 91,
                            "rating_id": 1,
                            "entity": {
                              "id": 33,
                              "entity_label": "Front Door Camera Angle",
                              "category": {
                                "id": 4,
                                "label": "Entrance",
                                "sort_order": 1
                              }
                            },
                            "rating": { "id": 1, "label": "Pass" },
                            "notes": [
                              {
                                "id": 501,
                                "camera_form_id": 210,
                                "note": "Slight glare; adjust exposure.",
                                "attachments": [
                                  {
                                    "id": 9001,
                                    "camera_form_note_id": 501,
                                    "path": "camera_forms/stores/12-brooklyn-atlantic-ave/2026-01-20/front-door-camera-angle-1.jpg",
                                    "url": "https://your-domain.tld/storage/camera_forms/stores/12-brooklyn-atlantic-ave/2026-01-20/front-door-camera-angle-1.jpg"
                                  }
                                ]
                              }
                            ]
                          }
                        ]
                      },
                      "errors": null
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedJson" },
          "403": { "$ref": "#/components/responses/ForbiddenJson" },
          "404": { "$ref": "#/components/responses/NotFoundJson" },
          "500": { "$ref": "#/components/responses/InternalServerErrorJson" }
        }
      },
      "put": {
        "tags": ["Camera Forms"],
        "summary": "Update an audit + associated camera forms/notes/attachments (logic preserved in your codebase)",
        "description": "Accepts multipart/form-data (images). Your controller indicates the update logic is preserved from the web version; the detailed behavior depends on your trimmed block.\n\nThe request validation supports:\n- entities.*.rating_id\n- entities.*.notes.*.id (existing note id)\n- entities.*.notes.*.remove_attachment_ids (remove attachments)\n- entities.*.remove_note_ids (remove notes)\n- entities.*.notes.*.images.* (new images)\n\nEven if you don't change anything, validation requires at least one entity to have a rating, note, image, or removal.",
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CameraFormUpdateRequest"
              },
              "examples": {
                "example_update_rating_and_remove_attachment": {
                  "value": {
                    "store_id": 12,
                    "date": "2026-01-21",
                    "entities": [
                      {
                        "entity_id": 33,
                        "rating_id": 2,
                        "notes": [
                          {
                            "id": 501,
                            "note": "Updated note text",
                            "remove_attachment_ids": [9001]
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Camera form updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiSuccessResponse" },
                "examples": {
                  "success": {
                    "value:": {
                      "status": "success",
                      "message": "Camera form updated successfully",
                      "data": {
                        "id": 91,
                        "store_id": 12,
                        "user_id": 5,
                        "date": "2026-01-21",
                        "camera_forms": [
                          {
                            "id": 210,
                            "entity_id": 33,
                            "rating_id": 2,
                            "notes": [
                              {
                                "id": 501,
                                "note": "Updated note text",
                                "attachments": []
                              }
                            ]
                          }
                        ]
                      },
                      "errors": null
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedJson" },
          "403": { "$ref": "#/components/responses/ForbiddenJson" },
          "404": { "$ref": "#/components/responses/NotFoundJson" },
          "422": { "$ref": "#/components/responses/ValidationErrorJson" },
          "500": {
            "description": "Failed to update camera form",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
                "examples": {
                  "failed": {
                    "value": {
                      "status": "error",
                      "message": "Failed to update camera form",
                      "data": null,
                      "errors": null
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Camera Forms"],
        "summary": "Delete an audit and all associated camera forms, notes, and attachments",
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Camera form deleted successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiSuccessResponse" },
                "examples": {
                  "success": {
                    "value": {
                      "status": "success",
                      "message": "Camera form deleted successfully",
                      "data": null,
                      "errors": null
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/UnauthorizedJson" },
          "403": { "$ref": "#/components/responses/ForbiddenJson" },
          "404": { "$ref": "#/components/responses/NotFoundJson" },
          "500": { "$ref": "#/components/responses/InternalServerErrorJson" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdPathParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 },
        "example": 91
      },
...
    },
    "responses": {
      "UnauthorizedJson": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "examples": {
              "unauthorized": {
                "value": {
                  "status": "error",
                  "message": "Unauthorized",
                  "data": null,
                  "errors": null
                }
              }
            }
          }
        }
      },
      "ForbiddenJson": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "examples": {
              "forbidden": {
                "value": {
                  "status": "error",
                  "message": "Forbidden",
                  "data": null,
                  "errors": null
                }
              }
            }
          }
        }
      },
      "NotFoundJson": {
        "description": "Not Found (Model not found / route not found)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LaravelNotFoundResponse"
            },
            "examples": {
              "model_not_found": {
                "value": {
                  "message": "No query results for model [App\\\\Models\\\\Audit] 999999"
                }
              }
            }
          }
        }
      },
      "ValidationErrorJson": {
        "description": "Validation error (Laravel 422)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LaravelValidationErrorResponse"
            },
            "examples": {
              "validation_failed": {
                "value": {
                  "message": "The given data was invalid.",
                  "errors": {
                    "store_id": ["The selected store id is invalid."],
                    "entities": [
                      "At least one entity must have a rating, note, or image."
                    ],
                    "entities.0.entity_id": [
                      "The selected entities.0.entity_id is invalid."
                    ],
                    "entities.0.notes.0.images.0": [
                      "The entities.0.notes.0.images.0 must be an image.",
                      "The entities.0.notes.0.images.0 may not be greater than 5120 kilobytes."
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "InternalServerErrorJson": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiErrorResponse" },
            "examples": {
              "server_error": {
                "value": {
                  "status": "error",
                  "message": "Server Error",
                  "data": null,
                  "errors": null
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "ApiSuccessResponse": {
        "type": "object",
        "required": ["status", "message", "data", "errors"],
        "properties": {
          "status": { "type": "string", "enum": ["success"] },
          "message": { "type": "string" },
          "data": {},
          "errors": { "type": ["object", "null"], "default": null }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "required": ["status", "message", "data", "errors"],
        "properties": {
          "status": { "type": "string", "enum": ["error"] },
          "message": { "type": "string" },
          "data": {
            "type": ["object", "array", "string", "number", "boolean", "null"],
            "default": null
          },
          "errors": {
            "type": ["object", "array", "string", "null"],
            "default": null
          }
        }
      },
      "LaravelValidationErrorResponse": {
        "type": "object",
        "required": ["message", "errors"],
        "properties": {
          "message": {
            "type": "string",
            "example": "The given data was invalid."
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            }
          }
        }
      },
      "LaravelNotFoundResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "CameraFormUpdateRequest": {
        "type": "object",
        "required": ["store_id", "date", "entities"],
        "properties": {
          "store_id": { "type": "integer", "example": 12 },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-01-21"
          },
          "entities": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/CameraFormEntityUpdateRow"
            }
          }
        }
      },
      "CameraFormEntityUpdateRow": {
        "type": "object",
        "required": ["entity_id"],
        "properties": {
          "entity_id": { "type": "integer", "example": 33 },
          "rating_id": { "type": ["integer", "null"], "example": 2 },
          "notes": {
            "type": ["array", "null"],
            "items": { "$ref": "#/components/schemas/CameraFormNoteUpdateRow" }
          },
          "remove_note_ids": {
            "type": ["array", "null"],
            "items": { "type": "integer" },
            "example": [501]
          }
        }
      },
      "CameraFormNoteUpdateRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": ["integer", "null"],
            "example": 501,
            "description": "Existing note id (optional)."
          },
          "note": {
            "type": ["string", "null"],
            "example": "Updated note text"
          },
          "images": {
            "type": ["array", "null"],
            "items": { "type": "string", "format": "binary" }
          },
          "remove_attachment_ids": {
            "type": ["array", "null"],
            "items": { "type": "integer" },
            "example": [9001]
          }
        }
      },
...
    }
  }
}